% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_message}
\alias{log_message}
\title{Print Diagnostic Message}
\usage{
log_message(..., verbose = T)
}
\arguments{
\item{...}{Text to print}

\item{verbose}{Logical. Whether to display messages}
}
\description{
Print Diagnostic Message
}
