% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{prettyrainbow}
\alias{prettyrainbow}
\title{Define Default Color Palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
prettyrainbow
}
\description{
Define Default Color Palette
}
\keyword{datasets}
