% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_fidelity.R
\name{score_fidelity}
\alias{score_fidelity}
\title{Estimate the organ fidelity of Human Intestinal Organoids}
\usage{
score_fidelity(
  que_obj,
  organ_ref_dir = "../Ref_data_for_projection_to_fetal_atlas/",
  k = 20,
  group_by = NULL,
  group_cols = NULL,
  organ_cols = setNames(c("#9D0142", "#5D4EA2", "#FCB163", "#1B7635", "#4ea7b0",
    "#696969"), c("Intestine", "Stomach", "Lung", "Esophagus", "Liver", "Pancreas")),
  return_object = TRUE,
  output_dir = "./",
  probability_file_name = "Plot_barplot_projected_developing_organ_probability.pdf",
  do_projection = TRUE,
  projection_file_name = "Plot_UMAP_RSS_projection_to_developing_reference.png",
  verbose = TRUE
)
}
\arguments{
\item{que_obj}{Seurat object in query}

\item{organ_ref_dir}{A character string to specify the path to the folder with the fetal
reference (Yu et al, Cell, 2021) downloaded from Mendeley data https://doi.org/10.17632/x53tts3zfr.
Default is \code{'../Ref_data_for_projection_to_fetal_atlas/'}}

\item{k}{Numeric. Specify the number of k-nearest neighboring cells in the reference atlas
for each query organoid cell for fidelity estimation.}

\item{group_by}{A character string to specify the name of the meta.data field of a combined
Seurat object from more than one groups. Only applicable when the \code{'que_obj'} is a
combination of different groups and you would like to separate them. Default is \code{'NULL'}}

\item{group_cols}{A vector to specify the color(s) of query group(s) for plotting. Names of
the vector specify the group, values specify colors in Hex format.}

\item{organ_cols}{A vector to specify the colors of organ identity in reference atlas. Names
of the vector specify the group, values specify colors in Hex format.}

\item{return_object}{Logical. Whether to return a Seurat object with inferred organ identity
for each query cell}

\item{output_dir}{A character string to specify output directory}

\item{probability_file_name}{A character string to specify the file name of the plot showing
the projected organ probability}

\item{do_projection}{Logical. Whether to project the query cells to the UMAP embedding of
the reference atlas for visualization}

\item{projection_file_name}{A character string to specify the file name of the plot projecting
query cells to the UMAP embedding of reference atlas}

\item{verbose}{Logical. Whether to print log message}
}
\value{
A bar plot to show the organ projection probability, a UMAP plot to show projected UMAP
embedding of the query cell on reference atlas and an updated Seurat object.\cr
The updated Seurat has three new components:
\itemize{
\item Meta.data field \code{'Mapped_fetal_organ'} indicating the inferred organ identity of each query cell;
\item Dimension reduction object \code{'rss'} representing similarity of each query cell to the fetal
similarity spectrum;
\item Dimension reduction object \code{'rss_umap'} representing projected UMAP embedding of each query cell
on the fetal reference atlas
}
}
\description{
This function estimates the organ fidelity of cells in pluripotent stem cell
derived human intestinal organoids by comparing to a developing human multi-
endodermal-organ cell atlas
}
