% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_maturity.R
\name{score_maturity}
\alias{score_maturity}
\title{Estimate the Stem Cell Maturity of Human Intestinal Organoids}
\usage{
score_maturity(
  que_obj,
  group_by = NULL,
  score_method = c("deconvolution_based", "similarity_based"),
  reference_age = "adult",
  add_landmark_data = TRUE,
  selected_landmark_age = c("thio-4-week", "fetal-7-week", "fetal-8-week", "adult"),
  group_cols = NULL,
  do_plot = TRUE,
  output_dir = "./",
  plot_name = "Plot_ISC_maturity.pdf",
  verbose = TRUE
)
}
\arguments{
\item{que_obj}{Seurat object in query}

\item{group_by}{A character string to specify the name of the meta.data field of a combined
Seurat object from more than one groups. Only applicable when the \code{'que_obj'} is a
combination of different groups and you would like to separate them. Default is \code{'NULL'}}

\item{score_method}{Specify the method to estimate stem cell maturity. Currently support
\code{'similarity_based'} and \code{'deconvolution_based'}. Default is \code{'deconvolution_based'}.}

\item{reference_age}{A character string to specify which age(s) as reference for transcriptome
similarity calculation. Currently support one or several values from \code{'thio-4-week'},
\code{'thio-8-week'}, \code{'fetal-7-week'}, \code{'fetal-8-week'}, \code{'fetal-10-week'},
\code{'fetal-11-week'}, \code{'fetal-12-week'}, \code{'fetal-14-week'}, \code{'fetal-17-week'},
\code{'fetal-18-week'}, \code{'fetal-19-week'}, \code{'adult'}. If choose multiple reference
ages, only median similarity to each reference for each query group is shown in the output plot.}

\item{add_landmark_data}{Logical. Whether to add reference ISC cellular data to the query set}

\item{selected_landmark_age}{A character string to specify which reference ISC cellular data
to be added in the query set}

\item{group_cols}{A vector to specify the color(s) of query group(s) for plotting. Names of
the vector specify the group, values specify colors in Hex format.}

\item{do_plot}{Logical. Whether to do plotting for result visualization}

\item{output_dir}{A character string to specify output directory}

\item{plot_name}{A character string to specify the file name of the plot}

\item{verbose}{Logical. Whether to print log message}
}
\value{
A plot to show the stem cell maturity and a list of numeric values of maturity
}
\description{
This function provides two approaches to estimate the maturity of stem cells in pluripotent stem cell
derived human intestinal organoids. One is to calculate transcriptome similarity (Spearman's correlation
coefficient) between query cells and reference intestinal stem cells (ISCs). The other is to decompose
the cellular transcriptome into spheroid and maturing fetal (19 post conceptional week (PCW)) ISC
components using quadratic programming and quantifies the maturity of query cells as the proportion
of 19 PCW fetal ISC components.
}
